Refactored some modules for the Boolean operations
