+++ plato.cabal
@@ -26,11 +26,12 @@ library
                       Tuura.Concept.FSM,
                       Tuura.Concept.FSM.Simulation,
                       Tuura.Concept.FSM.Translation,
-                      Tuura.Parser.Boolean,
+                      Tuura.Boolean.Parser,
+                      Tuura.Boolean.Options,
+                      Tuura.Boolean,
                       Tuura.Plato.Translate.Options,
                       Tuura.Plato.Translate.Translation,
-                      Tuura.Plato.BoolToConcept.Options,
-                      Tuura.Plato.BoolToConcept.BooleanFunctions
+                      Tuura.Plato.BoolToConcept
     other-modules:    Tuura.Concept.Abstract,
                       Tuura.Concept.Circuit,
                       Tuura.Concept.Circuit.Basic,
